Chore(persistent-data): Fix failing tests
