+++ .travis.yml
@@ -11,7 +11,6 @@ before_install:
 
 install:
  - pip install -r requirements.txt
- # Submit coverage report to coveralls.io.
  - pip install coveralls
 
 script:
