--- app/tests/__init__.py
@@ -1,11 +1,0 @@
- Volume in drive C has no label.
- Volume Serial Number is E439-29F7
-
- Directory of C:\Users\User\andela_35\todo-list-api\app\tests
-
-03-Oct-18  08:06 AM    <DIR>          .
-03-Oct-18  08:06 AM    <DIR>          ..
-03-Oct-18  08:06 AM             4,977 tests.py
-03-Oct-18  08:06 AM                 0 __init__.py
-               2 File(s)          4,977 bytes
-               2 Dir(s)  58,869,288,960 bytes free
